# openapi.json
{"openapi":"3.1.0","info":{"title":"ComfyUI API","description":"ComfyUI API服务，提供图像生成功能","version":"0.1.0"},"paths":{"/api/generate_image":{"post":{"tags":["Image Generation"],"summary":"Generate Image","description":"生成图像API\n\n接收图像生成参数，提交图像生成任务，并返回任务ID","operationId":"generate_image_api_generate_image_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageGenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/task_status/{task_id}":{"get":{"tags":["Image Generation"],"summary":"Get Task Status","description":"获取任务状态API\n\n根据任务ID获取图像生成任务的状态","operationId":"get_task_status_api_task_status__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/get_file/{prompt_id}":{"get":{"tags":["Image Generation"],"summary":"Get File","description":"获取图像文件API\n\n根据prompt_id获取生成的图像文件路径","operationId":"get_file_api_get_file__prompt_id__get","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"string","title":"Prompt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageGenerationRequest":{"properties":{"prompt":{"type":"string","title":"Prompt","default":""},"negative_prompt":{"type":"string","title":"Negative Prompt","default":""},"width":{"type":"integer","title":"Width","default":512},"height":{"type":"integer","title":"Height","default":512},"batch_size":{"type":"integer","title":"Batch Size","default":4}},"type":"object","title":"ImageGenerationRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}